ANDROID-49 экран ошибки перенесен в core-ui
